+++ Client/Ecosia/Migration/AppDelegate+Ecosia.swift
@@ -8,27 +8,6 @@ import Core
 extension AppDelegate {
     func migrateEcosiaContents() {
         guard EcosiaImport.isNeeded, let profile = profile else { return }
-
-        let ecosiaImport = EcosiaImport(profile: profile, tabManager: self.tabManager)
-        ecosiaImport.migrate(progress: { progress in
-            print("progress: \(progress) ")
-        }){ [weak self] migration in
-            if case .succeeded = migration.favorites,
-               case .succeeded = migration.tabs,
-               case .succeeded = migration.history {
-                self?.cleanUp()
-                Analytics.shared.migration(true)
-            } else {
-                Analytics.shared.migration(false)
-            }
-            
-            Core.User.shared.migrated = true
-        }
-    }
-    
-    private func cleanUp() {
-        History().deleteAll()
-        Favourites().items = []
-        Tabs().clear()
+        window?.rootViewController?.present(LoadingScreen(profile: profile, tabManager: tabManager), animated: false)
     }
 }
